UDP: Signficant refactoring and simplification.
This work is inspired by the DTLS work, and harmonizes the UDP implementation
with DTLS somewhat.

This should make it more resilient to failures, although there is no longer any
attempt to guard against sequencing (reorders, dupes) errors.  Applications that
need such protection should either add it themselves, or use a transport which
provides that guarantee (such as TCP).  Note that with devices and and such in
the way, such guarantees have never been perfect with SP anyway.

The UDP transport header sizes for this are now just 8 bytes (beyond the UDP header
itself.